requirements.txt
pprint
python-pptx
configparser